+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.1.0 - 2024-10-24
+
+### Added
+
+- Console command to clear block cache for entries related to expired entries
+
+### Documentation
+
+- Added examploe for CkEditor nested entries inside matrix blocks
+
 ## 3.0.0 - 2024-09-23
 
 ### Added